Σπάσιμο components σε υποmodules
